+++ .gitignore
@@ -1,44 +1,10 @@
 *.m3u
 *.png
-logs/*.txt
+logs/*
 *.json
 
-logs/debug_log_2023-11-24.txt
 temp/favicon.ico
-logs/debug_log_2023-11-24.txt
-logs/debug_log_2023-11-24.txt
-get-anime.py
-logs/debug_log_2023-11-24.txt
-logs/debug_log_2023-11-26.txt.2023-11-26
-logs/debug_log_2023-11-25.txt.2023-11-25
-logs/debug_log_2023-11-27.txt.2023-11-27
-poster_cache.txt
-helper/poster_utils.py
-helper/__pycache__/poster_utils.cpython-312.pyc
-legacy-1.8.0/config.ini
-legacy-1.8.0/LICENSE
-legacy-1.8.0/README.md
-legacy-1.8.0/requirements.txt
-logs/debug_log_2023-11-28.txt.2023-11-28
-logs/debug_log_2023-11-28.txt.2023-11-29
-logs/debug_log_2023-11-29.txt.2023-11-29
-logs/debug_log_2023-11-30.txt.2023-11-30
-logs/debug_log_2023-12-01.txt.2023-12-01
-logs/debug_log_2023-12-01.txt.2023-12-02
-logs/debug_log_2023-12-03.txt.2023-12-03
-logs/debug_log_2023-12-03.txt.2023-12-04
-logs/debug_log_2023-12-03.txt.2023-12-05
-logs/debug_log_2023-12-06.txt.2023-12-06
-logs/debug_log_2023-12-07.txt.2023-12-07
-logs/debug_log_2023-12-08.txt.2023-12-08
-logs/debug_log_2023-12-09.txt.2023-12-09
-shared/get_data.py
-utils/config.ini
-utils/file_config.py
-utils/file_playlist.py
-utils/file_utils.py
-utils/logging.py
-utils/__pycache__/file_config.cpython-312.pyc
-utils/__pycache__/file_playlist.cpython-312.pyc
-utils/__pycache__/file_utils.cpython-312.pyc
-utils/__pycache__/logging.cpython-312.pyc
+utils/__pycache__/*
+__pycache__/*
+gui/__pycache__/__init__.cpython-312.pyc
+gui/__pycache__/app.cpython-312.pyc
