+++ config.ini
@@ -2,8 +2,8 @@
 base_url = anilibria.tv
 api_version = v3
 stream_video_url = cache.libria.fun
-video_player_path = /Applications/VLC.app/Contents/MacOS/VLC
-torrent_client_path = /Applications/Transmission.app/Contents/MacOS/Transmission
+video_player_path = C:\Program Files\VideoLAN\VLC\vlc.exe
+torrent_client_path = C:\Program Files\Transmission\transmission-qt.exe
 [Logging]
 log_level = DEBUG
