+++ logging.conf
@@ -1,0 +1,36 @@
+# logging.conf
+
+[loggers]
+keys=root,AnimePlayerApp
+
+[handlers]
+keys=consoleHandler,fileHandler
+
+[formatters]
+keys=defaultFormatter
+
+[logger_root]
+level=NOTSET
+handlers=consoleHandler, fileHandler
+
+[logger_AnimePlayerApp]
+level=DEBUG
+handlers=consoleHandler,fileHandler
+qualname=AnimePlayerApp
+propagate=0
+
+[handler_consoleHandler]
+class=StreamHandler
+level=DEBUG
+formatter=defaultFormatter
+args=(sys.stdout,)
+
+[handler_fileHandler]
+class=logging.handlers.TimedRotatingFileHandler
+level=DEBUG
+formatter=defaultFormatter
+args=('logs/debug_log.txt', 'midnight', 1, 7, 'utf-8')
+
+[formatter_defaultFormatter]
+format=%(asctime)s - %(name)s - %(levelname)s - %(message)s
+datefmt=
